Move icons module from shared to components
