+++ src/app/components/components.lazy.ts
@@ -1,10 +1,16 @@
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
+  // App global components
   {
     path: '~',
     loadChildren: () => import('./masterpage/masterpage.module').then(m => m.MasterPageModule),
   },
+  {
+    path: '~',
+    loadChildren: () => import('./icons/icons.module').then(m => m.IconsModule),
+  },
+  // Angular libs
   {
     path: '~',
     loadChildren: () => import('@angular/forms').then(m => m.FormsModule),
@@ -13,9 +19,14 @@ const routes: Routes = [
     path: '~',
     loadChildren: () => import('@angular/forms').then(m => m.ReactiveFormsModule),
   },
+  // Prime libs
   {
     path: '~',
     loadChildren: () => import('primeng/card').then(m => m.CardModule),
+  },
+  {
+    path: '~',
+    loadChildren: () => import('primeng/messages').then(m => m.MessagesModule),
   },
   {
     path: '~',
