+++ src/app/components/icons/icons.module.ts
