+++ src/app/shared/index.ts
@@ -4,7 +4,6 @@ export * from './utils';
 export * from './stores';
 export * from './pipes';
 export * from './services';
-export * from './icons/icons.module';
 export * from './guards';
 export * from './interceptors';
 export * from './models';
